Group sockets by category
