+++ core/src/main/scala/io/fmq/socket/pubsub/Subscriber.scala
@@ -1,4 +1,5 @@
 package io.fmq.socket
+package pubsub
 
 import java.nio.charset.StandardCharsets
