+++ core/src/main/scala/io/fmq/socket/pubsub/XPublisher.scala
@@ -1,4 +1,4 @@
-package io.fmq.socket
+package io.fmq.socket.pubsub
 
 import cats.effect.{Blocker, ContextShift, Resource, Sync}
 import cats.syntax.functor._
@@ -7,7 +7,7 @@ import io.fmq.socket.api.{CommonOptions, SendOptions, SocketOptions}
 import io.fmq.socket.internal.Bind
 import org.zeromq.ZMQ
 
-final class Publisher[F[_]: ContextShift] private[fmq] (
+final class XPublisher[F[_]: ContextShift] private[fmq] (
     protected[fmq] val socket: ZMQ.Socket,
     blocker: Blocker
 )(implicit protected val F: Sync[F])
@@ -15,12 +15,12 @@ final class Publisher[F[_]: ContextShift] private[fmq] (
     with CommonOptions.All[F]
     with SendOptions.All[F] {
 
-  def bind[P <: Protocol, A <: Address: Complete[P, *]](uri: Uri[P, A]): Resource[F, ProducerSocket[F, P, A]] =
-    Bind.bind[F, P, A](uri, socket, blocker).as(ProducerSocket.create(socket, uri))
+  def bind[P <: Protocol, A <: Address: Complete[P, *]](uri: Uri[P, A]): Resource[F, XPublisherSocket[F, P, A]] =
+    Bind.bind[F, P, A](uri, socket, blocker).as(new XPublisherSocket(socket, uri))
 
-  def bindToRandomPort(uri: Uri.TCP[Address.HostOnly]): Resource[F, ProducerSocket[F, Protocol.TCP, Address.Full]] =
+  def bindToRandomPort(uri: Uri.TCP[Address.HostOnly]): Resource[F, XPublisherSocket[F, Protocol.TCP, Address.Full]] =
     for {
       uriFull <- Bind.bindToRandomPort[F](uri, socket, blocker)
-    } yield ProducerSocket.create(socket, uriFull)
+    } yield new XPublisherSocket(socket, uriFull)
 
 }
